refactored the code space, abstracted one more layer for all UI layer to use
