--- cli/oe/testdata/dsl.json
@@ -1,39 +1,0 @@
-[
-  {
-    "description": "Empty DSL",
-    "dsl": {
-      "API": null,
-      "Engine": {
-        "providers": null,
-        "provisioners": null,
-        "resources": null,
-        "schedule_order": null,
-        "solutions": [],
-        "systems": null,
-        "tools": null
-      },
-      "Provisioners": null,
-      "Resources": null,
-      "Systems": null,
-      "Tools": null
-    },
-    "tests": [
-      {
-        "description": "Create engine",
-        "action": "CreateEngine",
-        "want": [{}]
-      },
-      {
-        "description": "Run engine",
-        "action": "Run",
-        "args": ["create"],
-        "want": [null]
-      },
-      {
-        "description": "Get (no) solutions",
-        "action": "GetSolutions",
-        "want": [[]]
-      }
-    ]
-  }
-]
