+++ util/util.go
@@ -1,0 +1,15 @@
+// util is the util(util method) package for all of the openengine
+package util
+
+import (
+	"os"
+)
+
+func FileExists(filename string) bool {
+	info, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false
+	}
+
+	return !info.IsDir()
+}
